==> src/development/0_planning.ipynb <==
**1. Define the task you want the model to make?**

Predict the duration of taxi or limousine rides with information known at the start of the taxi or limousine ride.

**2. What is the input data?**

Pickup time, duration, passengers, rate code, store and forward, location ids.

The fields we can use for analysis but not directly for prediction are: trip distance, paymen ttype, payment amounts, whether a toll was used.

Primary inputs are traffic at the time of day and the start/end locations.

**3. What output will you have?**

The predicted duration.

**4. Is this regression, classification, or another type?**

It is regression.

**5. What is your performance metric?**

MSE is our performance metric.  A cap to ensure all targets are between 0 minutes and 30 minutes.  So if it falls below 0 it becomes 0 and if it goes above 30 minutes it's capped at 30 minutes.

**6. What is your finish line performance metric value? (To know when you're results
are good enough to conclude the project)**

We know the results are good enough if it predicts with +/- 2 minutes of accuracy on average.
