==> src/development/1.1_data_extraction.ipynb <==
import gc
import numpy as np
import pandas as pd
from backroom import *
df = pd.read_csv("../../data/raw/taxi_data.csv")
df.rename(columns={'Unnamed: 0': 'id'}, inplace=True)
for column in ['tpep_pickup_datetime','tpep_dropoff_datetime']:
    df[column] = pd.to_datetime(df[column])
df['is_store_and_fwd'] = df['store_and_fwd_flag'].apply(lambda x: 1 if x == 'Y' else 0)
df['pickup_day_of_week'] = df['tpep_pickup_datetime'].apply(lambda x: x.dayofweek)
df['pickup_time_of_day'] = df['tpep_pickup_datetime'].apply(lambda x: '00-03' if x.hour < 4 else '04-07' if x.hour < 8 else '08-11' if x.hour < 12 else '12-15' if x.hour < 16 else '16-19' if x.hour < 20 else '20-23')
df['duration_diff'] = df['tpep_dropoff_datetime'] - df['tpep_pickup_datetime']
df['duration_seconds'] = df['duration_diff'].dt.total_seconds()
df.drop(['store_and_fwd_flag', 'duration_diff'], axis='columns', inplace=True)
df.to_csv("../../data/interim/df.csv", index=False)
free_up_memory(locals())
